requirements.txt
django
djangorestframework
channels
daphne
ffmpeg-python
requests
python-dotenv
openai-whisper
transformers
torch

